refactor: modularize CI summary
